Restructure and rename modules (#101)
* Updates in EntryStore

- update p2panda-rs
- restructure modules
- introduce new replication methods

* Revert p2panda-rs version and remove new methods in EntryStore

* Remove JSON RPC services

* Remove unused JSON RPC test helpers

* Update CHANGELOG

* Add docstrings
