+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Service manager [#90](https://github.com/p2panda/aquadoggo/pull/90)
 - Service error handling, refactor runtime [#92](https://github.com/p2panda/aquadoggo/pull/92)
 - Refactor module structure, propagate errors in worker to service manager [#97](https://github.com/p2panda/aquadoggo/pull/97)
+- Restructure storage modules and remove JSON RPC [#101](https://github.com/p2panda/aquadoggo/pull/101)
 
 ### Changed
