--- aquadoggo/src/rpc/methods/mod.rs
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-mod entry_args;
-mod publish_entry;
-mod query_entries;
-
-pub use entry_args::get_entry_args;
-pub use publish_entry::publish_entry;
-pub use query_entries::query_entries;
