+++ aquadoggo/src/rpc/mod.rs
@@ -1,12 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-mod api;
-mod methods;
 mod request;
 mod response;
-mod server;
 
-pub use api::{build_rpc_api_service, RpcApiService, RpcApiState};
 pub use request::{EntryArgsRequest, PublishEntryRequest};
 pub use response::{EntryArgsResponse, PublishEntryResponse};
-pub use server::{handle_get_http_request, handle_http_request};
