+++ aquadoggo/src/test_helpers.rs
@@ -155,60 +155,3 @@ pub fn random_entry_hash() -> String {
         .as_str()
         .to_owned()
 }
-
-// Helper method to generate valid JSON RPC request string
-pub(crate) fn rpc_request(method: &str, params: &str) -> String {
-    format!(
-        r#"{{
-            "jsonrpc": "2.0",
-            "method": "{}",
-            "params": {},
-            "id": 1
-        }}"#,
-        method, params
-    )
-    .replace(" ", "")
-    .replace("\n", "")
-}
-
-// Helper method to generate valid JSON RPC response string
-pub(crate) fn rpc_response(result: &str) -> String {
-    format!(
-        r#"{{
-            "jsonrpc": "2.0",
-            "result": {},
-            "id": 1
-        }}"#,
-        result
-    )
-    .replace(" ", "")
-    .replace("\n", "")
-}
-
-// Helper method to generate valid JSON RPC error response string
-pub(crate) fn rpc_error(message: &str) -> String {
-    format!(
-        r#"{{
-            "jsonrpc": "2.0",
-            "error": {{
-                "code": 0,
-                "message": "<message>"
-            }},
-            "id": 1
-        }}"#
-    )
-    .replace(" ", "")
-    .replace("\n", "")
-    .replace("<message>", message)
-}
-
-// Helper method to handle JSON RPC HTTP request and return response
-pub(crate) async fn handle_http(client: &TestClient, request: String) -> String {
-    let response = client
-        .post("/")
-        .body(request)
-        .header("content-type", HeaderValue::from_static("application/json"))
-        .send()
-        .await;
-    response.text().await
-}
